Uploaded the complete project.
added 6 files, 4 html, 1 css, 1 svg Medallia logo.
The project is a simple login form which will show some password validation. The login form will work if you enter any name and password of more than 8 characters but the next page will show "Login failed";however if you enter the Username:- test1 and Password:- test12345678 the next page will show the login success.
A signup page which will have certain fields to be filled and marked as required and when submitted will show the data on the next page. This also have a rest button to clear the  data entered on the signup form.